Add weather API hooks
