+++ package.json
@@ -8,7 +8,8 @@
     "@fortawesome/free-regular-svg-icons": "^5.15.0",
     "@fortawesome/free-solid-svg-icons": "^5.15.0",
     "react": "^16.13.1",
-    "react-dom": "^16.13.1"
+    "react-dom": "^16.13.1",
+    "react-query": "^2.23.0"
   },
   "scripts": {
     "start": "react-scripts start",
