+++ src/common/constants.ts
@@ -1,1 +1,3 @@
 export const APP_NAME = "Weather demo";
+export const CONTACT_EMAIL = "wm.wooley@gmail.com";
+export const URL = "https://github.com/michaelwooley/frontend-demo-react";
