+++ src/services/weather.ts
@@ -1,75 +1,70 @@
-type WeatherApiContext = (string | Record<string, unknown>)[];
+import { CONTACT_EMAIL, URL } from "common/constants";
+import {
+  WeatherApiStationObservationLatest,
+  WeatherApiStationObservations,
+} from "types/weather.types";
+import { useQuery } from "react-query";
 
-type WeatherStat<V extends number | null = number> = {
-  value: number;
-  unitCode: string;
-  qualityControl: null | string;
+const BASE_URL = "https://api.weather.gov";
+
+var HEADERS = new Headers();
+HEADERS.append("Accept", "application/geo+json");
+HEADERS.append("User-Agent", `(${URL},${CONTACT_EMAIL})`);
+HEADERS.append("Origin", URL);
+
+const REQUEST_INIT = {
+  method: "GET",
+  headers: HEADERS,
+} as RequestInit;
+
+/**
+ * Common weather API request
+ *
+ * NOTE Cannot request https resources from localhost in firefox. See https://developer.mozilla.org/en-US/docs/Web/HTTP/CORS/Errors/CORSDidNotSucceed?utm_source=devtools&utm_medium=firefox-cors-errors&utm_campaign=default
+ *
+ * @param url
+ */
+const _callWeatherStation = async <T>(url: string): Promise<T> => {
+  const res = await fetch(url, REQUEST_INIT);
+
+  if (!res.ok) {
+    const r = await res.json();
+    throw new Error(r.detail || "Station not found");
+  }
+
+  return await res.json();
 };
 
-export interface StationObservation {
-  id: string;
-  type: "Feature";
-  geometry: {
-    type: "Point";
-    coordinates: number[];
-  };
-  properties: {
-    "@id": string;
-    "@type": string;
-    elevation: {
-      value: 382;
-      unitCode: string;
-    };
-    station: string;
-    timestamp: string;
-    rawMessage: string;
-    textDescription: string;
-    icon: string;
-    presentWeather: [];
-    temperature: WeatherStat;
-    dewpoint: WeatherStat;
-    windDirection: WeatherStat;
-    windSpeed: WeatherStat;
-    windGust: WeatherStat<number | null>;
-    barometricPressure: WeatherStat;
-    seaLevelPressure: WeatherStat<number | null>;
-    visibility: WeatherStat;
-    maxTemperatureLast24Hours: WeatherStat<number | null>;
-    minTemperatureLast24Hours: WeatherStat<number | null>;
-    precipitationLastHour: {
-      value: number | null;
-      unitCode: string;
-      qualityControl: string;
-    };
-    precipitationLast3Hours: WeatherStat<number | null>;
-    precipitationLast6Hours: WeatherStat<number | null>;
-    relativeHumidity: WeatherStat;
-    windChill: WeatherStat<number | null>;
-    heatIndex: WeatherStat;
-    cloudLayers: [
-      {
-        base: {
-          value: number | null;
-          unitCode: string;
-        };
-        amount: string;
-      }
-    ];
-  };
+export const weatherLatest = async (
+  _: string,
+  stationId: string
+): Promise<WeatherApiStationObservationLatest> => {
+  const url = `${BASE_URL}/stations/${stationId}/observations/latest`;
+
+  return _callWeatherStation<WeatherApiStationObservationLatest>(url);
+};
+
+export const weatherHistorical = async (
+  _: string,
+  stationId: string,
+  limit = 50
+): Promise<WeatherApiStationObservations> => {
+  const url = `${BASE_URL}/stations/${stationId}/observations?limit=${limit}`;
+
+  const data = await _callWeatherStation<WeatherApiStationObservations>(url);
+
+  // Will return "ok" even if stationId is bad
+  if (data.features.length === 0) {
+    throw new Error("Station not found");
+  }
+
+  return data;
+};
+
+export function useWeatherLatest(stationId: string) {
+  return useQuery(["weather-latest", stationId], weatherLatest);
 }
 
-/**
- * Response from: https://api.weather.gov/stations/:station/observations/latest
- */
-export interface WeatherApiStationObservationLatest extends StationObservation {
-  context: WeatherApiContext;
-}
-
-/**
- * Response from: https://api.weather.gov/stations/:station/observations?start={Date}&end={Date}&limit={number:50}
- */
-export interface WeatherApiStationObservations {
-  context: WeatherApiContext;
-  type: "FeatureCollection";
-  features: StationObservation[];
+export function useWeatherHistorical(stationId: string, limit = 50) {
+  return useQuery(["weather-historical", stationId, limit], weatherHistorical);
 }
